Split off param_core from parameters
